🚚 Rename: sidebar 폴더 생성
